separé en modulos y modifique add() para que utilize el método isEqual()
