+++ index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+	<meta charset="UTF-8">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<meta http-equiv="X-UA-Compatible" content="ie=edge">
+	<title>Kindle Bueno</title>
+</head>
+<body>
+	
+	<h1>Hola medialuneros++!</h1>
+
+	<script type="module" src="app.js"></script>
+</body>
+</html>
